// apps/frontend/types/trading.ts
export interface Stock {
  id: string
  symbol: string
  name: string
  instrumentId: string
  segment?: string
  strikePrice?: number
  optionType?: string
  expiry?: string
  lotSize?: number
}